requirements_colab.txt
mahotas
albumentations
segmentation_models_pytorch
zarr
pytorch-toolbelt
torch-lr-finder